+++ app/admin/dashboard.rb
@@ -1,3 +1,4 @@
 ActiveAdmin.register_page "Dashboard" do
-  
+  menu false
+
 end