+++ app/jobs/app_uninstalled_job.rb
@@ -1,4 +1,5 @@
 require "slack_notify"
+require "active_campaign_logger"
 
 class AppUninstalledJob < ApplicationJob
 
@@ -7,7 +8,16 @@ class AppUninstalledJob < ApplicationJob
     shop.with_shopify_session do
       shop.subscriptions.each { |s| s.destroy }
       shop.update_attributes(credit: 0, uninstalled_at: Time.now)
-      SlackNotify.uninstall(shop.domain)
+      slack_msg = "A shop has uninstalled Touchcard: #{shop.domain}."
+      SlackNotify.message(slack_msg)
+
+      sync_params = {
+          "email" => shop.email,
+          "tags" => "uninstalled"
+      }
+      sync_params.reject!{ |k,v| v.nil?}  # remove keys with no value
+      result = ActiveCampaign::client.contact_sync(sync_params)
+      ActiveCampaignLogger.log(sync_params, result)
     end
   end